requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
